Backend/requirements.txt
asgiref==3.10.0
Django==4.2.25
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==11.3.0
sqlparse==0.5.3
typing_extensions==4.15.0
requests>=2.25.0
django-cors-headers>=3.7.0
oauthlib>=3.1.0
requests-oauthlib>=1.3.0
google-generativeai>=0.3.0
python-decouple>=3.8